+++ PolynomialRegressionGeneric.h
@@ -45,7 +45,7 @@ public:
     PolynomialRegressionGeneric();
     
     // inherited as an abstract function from the base class
-    void generateCurve(unsigned int degree = 1) throw (CurveFittingException);
+    void generateCurve(size_t degree = 1) throw (CurveFittingException);
 
 };
